Move BytesToBits to bitmap_ops.h, clean up includes
